Add --verbose flag to the ci=duet-commmit alias
The --verbose flag will show your staged commit when you are editing
your commit message so that you know what you are committing,
thus can write a better commit message.
It does this in such a way that the staged commit does not end up as
part of the commit message (i.e. it's completely harmless).