`install.sh` failure opens log file for review
Previously failure of the overnight install scripts would open the Pickle Cat
(<https://dn.ht/picklecat>), a thoroughly repulsive website which had no
informational value other signaling the script had failed.

Not only was the picklecat an obscure signal, but it was also an
energy-intensive one: on an iMac17,1 with an Intel i7 4GHz 6700K with
91W TDP (thermal dissipation), it would chew up a 1½ (25% + 13%) cores.
Anecdotally, it was so CPU-intensive would spin up the fan on Greg's
laptop.

Opening the log file instead provides more information and helps save
the environment.

Co-authored-by: Reid Mitchell <rmitchell@pivotal.io>
Co-authored-by: Brian Cunnie <bcunnie@pivotal.io>