Set SKIP_SSL_VALIDATION to true after running int
We rely on the CLI `set_int_test_lite.sh` script to set up environment
variables when we run CLI integration tests. That script should set up
the necessary certificates to connect to an environment securely and
explicitly disables SKIP_SSL_VALIDATION at the end.

For some reason the certificate configuration does not seem to work with
our workstations. Since we also have no requirement for a secure
connection, this commit disables SSL validation when connecting to a
targeted environment.

Authored-by: Belinda Liu <bliu@pivotal.io>